Issue #22: Fix crash caused by null country code
I was getting reports of a NullPointerException here in the field.
My guess is that some international users have their Nest account
data populated differently. Or, maybe it is possible to not have
a country specified in your Nest account? Not sure how this can
be null, but either way this is the right fix.